requirements.txt
Django==4.2.7
Pillow==10.1.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-decouple==3.8
pytesseract==0.3.10
opencv-python==4.8.1.78
pandas==2.1.4
openpyxl==3.1.2
python-docx==1.1.0
python-pptx==0.6.23
celery==5.3.4
redis==5.0.1
whitenoise==6.6.0
gunicorn==21.2.0
